requirements.txt
fastapi==0.129.0
pymongo==4.16.0
python-dotenv==1.2.1
pwdlib[argon2,bcrypt]==0.3.0
python-jose[cryptography]==3.5.0
python-multipart==0.0.22
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.0.0
httpx==0.28.1
pydantic==2.12.5
pydantic_core==2.41.5
pydantic_settings==2.12.0
email-validator==2.3.0
uvicorn==0.40.0
slowapi==0.1.9
redis==7.1.1
pre-commit==4.5.1
ruff==0.14.13
google-genai==1.63.0
structlog==25.5.0

# Monitoring & Observability
opentelemetry-api==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-exporter-otlp==1.39.1
opentelemetry-instrumentation-fastapi==0.60b1
opentelemetry-instrumentation-pymongo==0.60b1
opentelemetry-instrumentation-logging==0.60b1
prometheus-fastapi-instrumentator==7.1.0
